core: ardupilot-manager: Improve board start/stop logic
- Deprecate `_current_platform` in favor of `_running_board`, making this the central information regarding the autopilot
- Merge `start_board` and `run_with_board` logic into `start_ardupilot` and `get_board_to_be_used`  methods
- Use same logic for start of all flight controllers, including SITL
- Adapt other methods accordingly
- Deprecates `/platform` methods on API. Fetching running board is now done on `/running_board` and changing it is a job for `/preferred_board`.

With the new logic, `_running_board` should not be changed to control which board to use. It just indicates which one is running. Controlling which board to use is now a job for `set_preferred_board`.

+++ core/services/ardupilot_manager/main.py
@@ -5,7 +5,7 @@ import logging
 import os
 import shutil
 from pathlib import Path
-from typing import Any, Dict, List, Set
+from typing import Any, Dict, List, Optional, Set
 
 from commonwealth.utils.apis import PrettyJSONResponse
 from commonwealth.utils.logs import InterceptHandler
@@ -86,7 +86,9 @@ def update_endpoints(endpoints: Set[Endpoint] = Body(...)) -> Any:
 @version(1, 0)
 def get_available_firmwares(response: Response, vehicle: Vehicle) -> Any:
     try:
-        return autopilot.get_available_firmwares(vehicle, autopilot.current_platform)
+        if not autopilot._running_board:
+            raise ValueError("No board running. Cannot get available firmwares.")
+        return autopilot.get_available_firmwares(vehicle, autopilot._running_board.platform)
     except Exception as error:
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         return {"message": f"{error}"}
@@ -96,8 +98,10 @@ def get_available_firmwares(response: Response, vehicle: Vehicle) -> Any:
 @version(1, 0)
 async def install_firmware_from_url(response: Response, url: str) -> Any:
     try:
+        if not autopilot._running_board:
+            raise ValueError("No board running. Cannot install firmware.")
         await autopilot.kill_ardupilot()
-        autopilot.install_firmware_from_url(url, autopilot.current_platform)
+        autopilot.install_firmware_from_url(url, autopilot._running_board.platform)
     except Exception as error:
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         return {"message": f"{error}"}
@@ -110,10 +114,12 @@ async def install_firmware_from_url(response: Response, url: str) -> Any:
 async def install_firmware_from_file(response: Response, binary: UploadFile = File(...)) -> Any:
     custom_firmware = Path.joinpath(autopilot.settings.firmware_folder, "custom_firmware")
     try:
+        if not autopilot._running_board:
+            raise ValueError("No board running. Cannot install firmware.")
         with open(custom_firmware, "wb") as buffer:
             shutil.copyfileobj(binary.file, buffer)
         await autopilot.kill_ardupilot()
-        autopilot.install_firmware_from_file(custom_firmware, autopilot.current_platform)
+        autopilot.install_firmware_from_file(custom_firmware, autopilot._running_board.platform)
         os.remove(custom_firmware)
     except InvalidFirmwareFile as error:
         response.status_code = status.HTTP_415_UNSUPPORTED_MEDIA_TYPE
@@ -126,29 +132,11 @@ async def install_firmware_from_file(response: Response, binary: UploadFile = Fi
         await autopilot.start_ardupilot()
 
 
-@app.get("/platform", response_model=Platform, summary="Check what is the current running platform.")
+@app.get("/running_board", response_model=Optional[FlightController], summary="Check which is the running board.")
 @version(1, 0)
-def platform(response: Response) -> Any:
+def board(response: Response) -> Any:
     try:
-        return autopilot.current_platform
-    except Exception as error:
-        response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
-        return {"message": f"{error}"}
-
-
-@app.post("/platform", summary="Toggle between SITL and default platform (auto-detected).")
-@version(1, 0)
-async def set_platform(response: Response, use_sitl: bool, sitl_frame: SITLFrame = SITLFrame.VECTORED) -> Any:
-    try:
-        if use_sitl:
-            autopilot.current_platform = Platform.SITL
-            autopilot.current_sitl_frame = sitl_frame
-        else:
-            autopilot.current_platform = None
-        logger.debug("Restarting ardupilot...")
-        await autopilot.kill_ardupilot()
-        await autopilot.start_ardupilot()
-        logger.debug("Ardupilot successfully restarted.")
+        return autopilot._running_board
     except Exception as error:
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         return {"message": f"{error}"}
@@ -194,8 +182,10 @@ async def stop(response: Response) -> Any:
 @version(1, 0)
 async def restore_default_firmware(response: Response) -> Any:
     try:
+        if not autopilot._running_board:
+            raise ValueError("No board running. Cannot restore default firmware.")
         await autopilot.kill_ardupilot()
-        autopilot.restore_default_firmware(autopilot.current_platform)
+        autopilot.restore_default_firmware(autopilot._running_board.platform)
     except Exception as error:
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         return {"message": f"{error}"}
@@ -238,9 +228,6 @@ app.mount("/", StaticFiles(directory=str(FRONTEND_FOLDER), html=True))
 
 
 if __name__ == "__main__":
-    if args.sitl:
-        autopilot.current_platform = Platform.SITL
-
     loop = asyncio.new_event_loop()
 
     # # Running uvicorn with log disabled so loguru can handle it
